Atmosphere

Helios.absolute_airmass — Method

absolute_airmass(relative_airmass, pressure)

Computes the absolute airmass [7].

\[ m_{\text{air}} = m_{\text{air, rel}} \frac{P} {P_{\text{atm}}} \]

where $m_{\text{air}}$ is the absolute airmass, $m_{\text{air, rel}}$ is the relative_airmass, $P$ is the pressure, and $P_{\text{atm}}$ is the atmospheric pressure.

Helios.altitude2pressure — Method

altitude2pressure(altitude)

Computes the site pressure from local altitude.

It follows the same assumption of pressure2altitude an it is calculated as

\[ P = \left(\frac{44331.514 - A}{11880.516}\right)^{ \frac{1}{0.1902632}}\]

where $P$ is the pressure and $A$ is the altitude.

Helios.aod_bb_hulstrom1980 — Method

aod_bb_hulstrom1980(aod380, aod500)

Computes the approximate broadband aerosol optical depth (AOD), as proposed in [8, 9].

It correlates broadband AOD using two wavelengths, 380 nm and 500 nm, denoted with aod380 and aod500, respectively.

Helios.apparent_vapor_scale_height_gueymard94 — Method

apparent_vapor_scale_height_gueymard94(air_temperature)

Computes the apparent water vapor scale height [km] as proposed in [10, Eq. 4].

\[ \tilde{H}_{\text{wv}} = 0.4976 + 1.5265\frac{T_{\text{air}} + T_0}{T_0} + \exp(13.6897\frac{T_{\text{air}} + T_0}{T_0} - 14.9188\frac{T_{\text{air}} + T_0}{T_0}^3)\]

$T_{\text{air}}$ is the air_temperature [°C]

$T_0$ is the absolute zero [°C]

Helios.linke_turbidity_kasten1996 — Method

linke_turbidity_kasten1996(absolute_airmass, precipitable_water, aod_bb)

Computes the Linke turbidity using Kasten pyrheliometric formula.

Note that broadband aerosol optical depth (AOD) can be approximated by AOD measured at 700 nm according to Molineaux [11]. Bird and Hulstrom offer an alternate approximation using broadband aerosol optical depth (aod_bb), measured at 380 nm and 500 nm. Based on original implementation by Armel Oumbe. Note that these calculations are only valid for absolute_airmass less than 5 and precipitable_water less than 5 cm.

Helios.linke_turbidity_meteotest — Method

linke_turbidity_meteotest(observer_latitude, observer_longitude, month::Int)

Computes the Linke turbidity based on SoDa [12, 13].

It performs a bilinear interpolation of the Linke turbidity array, for the corrsponding observer_latitude, observer_longitude, and month values. In such an array, rows represent global latitudes, columns represent global longitudes, and depth (third dimension) represents months of the year.

Helios.precipitable_water_gueymard94 — Method

precipitable_water_gueymard94(
    saturation_vapor_pressure,
    surface_vapor_density
)

Computes precipitable water [cm] as proposed in [10, Eq. 1].

The accuracy of this method is approximately 20% for moderate precipitable water (1-3 cm) and less accurate otherwise.

\[ w_{\text{prec}} = 0.1 \rho_{\text{surf,vap}} P_{\text{sat,vap}}\]

$\rho_{\text{surf,vap}}$ is the surface_vapor_density [km], see surface_vapor_density_gueymard94

$P_{\text{sat,vap}}$ is the saturation water vapor pressure [mbar], see saturation_vapor_pressure_oke2018, saturation_vapor_pressure_gueymard93

Helios.pressure2altitude — Method

pressure2altitude(pressure)

Computes the local altitude from site pressure.

The following assumptions [14] hold:

Base pressure: 101325 Pa

Temperature at zero altitude: 288.15 K

Gravitational acceleration: 9.80665 m/s^2

Lapse rate: -6.5E-3 K/m

Gas constant for air: 287.053 J/(kg K)

Relative Humidity: 0%

\[ A = 44331.5 - 4946.62\left(\frac{P}{100}\right)^{0.190263}\]

where $A$ is the altitude and $P$ is the pressure.

Helios.relative_airmass_gueymard1993 — Method

relative_airmass_gueymard1993(apparent_zenith)
relative_airmass_gueymard1993(s::SolarPosition)

Computes the relative airmass as proposed in [7, 15].

\[ m_{\text{air, rel}} = \frac{1}{\cos \tilde{\zeta}_{\text{top}} + 0.00176759 \tilde{\zeta}_{\text{top}}((94.37515 - \tilde{\zeta}_{\text{top}})^{-1.21563})}\]

where $\tilde{\zeta}_{\text{top}}$ is the Sun's apparent zenith, i.e., the complementary to 90° of the topocentric_apparent_elevation.

Helios.relative_airmass_gueymard2003 — Method

relative_airmass_gueymard2003(apparent_zenith)
relative_airmass_gueymard2003(s::SolarPosition)

Computes the relative airmass as proposed in [16, 17].

\[ m_{\text{air, rel}} = \frac{1}{\cos \tilde{\zeta}_{\text{top}} + 0.48353 \tilde{\zeta}_{\text{top}}^{0.095846} / (96.741 - \tilde{\zeta}_{\text{top}})^{1.754}}\]

where $\tilde{\zeta}_{\text{top}}$ is the Sun's apparent zenith, i.e., the complementary to 90° of the topocentric_apparent_elevation.

Helios.relative_airmass_kasten1966 — Method

relative_airmass_kasten1966(apparent_zenith)
relative_airmass_kasten1966(s::SolarPosition)

Computes the relative airmass as proposed in [18].

\[ m_{\text{air, rel}} = \frac{1}{\cos \tilde{\zeta}_{\text{top}} + 0.15(93.885 - \tilde{\zeta}_{\text{top}})^{-1.253})}\]

where $\tilde{\zeta}_{\text{top}}$ is the Sun's apparentzenith zenith, i.e., the complementary to 90° of the [`topocentricapparent_elevation`](@ref).

Helios.relative_airmass_kastenyoung1989 — Method

relative_arimass_kastenyoung1989(apparent_zenith)
relative_airmass_kastenyoung1989(s::SolarPosition)

Computes the relative airmass as proposed in [19].

\[ m_{\text{air, rel}} = \frac{1}{\cosd \tilde{\zeta}_{\text{top}} + 0.50572((96.07995 - \tilde{\zeta}_{\text{top}})^{- 1.6364})}\]

where $\tilde{\zeta}_{\text{top}}$ is the Sun's apparent zenith, i.e., the complementary to 90° of the topocentric_apparent_elevation.

Helios.relative_airmass_pickering2002 — Method

relative_airmass_pickering2002(apparent_zenith)
relative_airmass_pickering2002(s::SolarPosition)

Computes the relative airmass as proposed in [20].

\[ m_{\text{air, rel}} = \frac{1}{\sin(90 - \tilde{\zeta}_{\text{top}} + \frac{244}{165 + 47(90 - \tilde{\zeta}_{\text{top}})^{1.1}})\]

where $\tilde{\zeta}_{\text{top}}$ is the Sun's apparent zenith, i.e., the complementary to 90° of the topocentric_apparent_elevation.

Helios.relative_airmass_simple — Method

relative_airmass_simple(apparent_zenith)
relative_airmass_simple(s::SolarPosition)

Computes the relative airmass as secant of the Sun's apparent zenith.

\[m_{\text{air, rel}} = \sec \tilde{\zeta}_{\text{top}}\]

where $m_{\text{air, rel}}$ is the relative airmass and $\tilde{\zeta}_{\text{top}}$ is the Sun's apparent zenith, i.e., the complementary to 90° of the topocentric_apparent_elevation.

Helios.relative_airmass_young1994 — Method

relative_airmass_young1994(zenith)
relative_airmass_young1994(s::SolarPosition)

Computes the relative airmass as proposed in [21].

\[ m_{\text{air, rel}} = \frac{1.002432\cos^2 \zeta_{\text{top}} + 0.148386\cos \zeta_{\text{top}} + 0.0096467}{\cos^3 \zeta_{\text{top}} + 0.149864\cos^2 \zeta_{\text{top}} + 0.0102963\cos \zeta_{\text{top}} + 0.000303978}\]

where $\zeta_{\text{top}}$ is the Sun's zenith, i.e., the complementary to 90° of the topocentric_elevation.

Helios.relative_airmass_youngirvine1967 — Method

relative_airmass_youngirvine1967(zenith)
relative_airmass_youngirvine1967(s::SolarPosition)

Computes the relative airmass as proposed in [22].

\[ m_{\text{air, rel}} = \sec \zeta_{\text{top}} (1 - 0.0012( \sec \zeta_{\text{top}}^2 - 1))\]

where $\zeta_{\text{top}}$ is the Sun's zenith, i.e., the complementary to 90° of the topocentric_elevation.

Helios.relative_humidity_oke2018 — Method

relative_humidity_oke2018(
    vapor_pressure,
    saturation_vapor_pressure
)

Computes the relative humidity [%] as proposed in [23].

\[ u_{\text{rel}} = 100 \frac{P_{\text{sat,vap}}}{P_{\text{vap}}}\]

$P_{\text{sat,vap}}$ is the saturation water vapor pressure [mbar], see saturation_vapor_pressure_oke2018, saturation_vapor_pressure_gueymard93

$P_{\text{vap}}$ is the water vapor pressure [mbar], see vapor_pressure_oke2018

Helios.saturation_vapor_pressure_gueymard93 — Method

saturation_vapor_pressure_gueymard93(air_temperature)

Computes the saturation water vapor pressure [mbar] as proposed in [7, Eq. 1].

\[ P_{\text{sat,vap}} = \exp\left(22.330 - 49.140 \frac{100}{T_{\text{air}} + T_0} - 10.922 \left(\frac{100}{T_{\text{air}} + T_0}\right)^2 - 0.39015 \frac{T_{\text{air}} + T_0}{100} \right)\]

$T_{\text{air}}$ is the air_temperature [°C].

$T_0$ is the absolute zero [°C].

Helios.saturation_vapor_pressure_oke2018 — Method

saturation_vapor_pressure_oke2018(dew_temperature)

Computes the saturation water vapor pressure [mbar] as proposed in [23].

\[ P_{\text{sat,vap}} = 6.112\exp(17.62T_{\text{dew}} / (243.12 + T_{\text{dew}}))\]

where $T_{\text{dew}}$ is the dew_temperature [°C].

Helios.surface_vapor_density_gueymard94 — Method

surface_vapor_density_gueymard94(
    air_temperature,
    relative_humidity,
    apparent_vapor_scale_height
)

Computes the surface water vapor density [g/m^3] as proposed in [10, Eq. 2].

\[ \rho_{\text{surf,vap}} = \frac{216.7 u_{\text{rel}} \tilde{H}_{\text{wv}}}{T_{\text{air}}}\]

$u_{\text{rel}}$ is the relative_humidity [%]

$\tilde{H}_{\text{wv}}$ is the apparent_vapor_scale [km], see apparent_vapor_scale_height_gueymard94

$T_{\text{air}}$ is the air_temperature [°C]

Helios.vapor_pressure_oke2018 — Method

vapor_pressure_oke2018(air_temperature)

Computes the water vapor pressure [mbar] as proposed in [23].

\[ P_{\text{vap}} = 6.112\exp\left( \frac{17.62T_{\text{air}}}{243.12 + T_{\text{air}}})\right)\]

where $T_{\text{air}}$ is the air_temperature [°C].