Clear-sky

The following functions implement different models to calculate the irradiance components under the clear-sky assumption.

Helios.clearsky_haurwitz — Method

clearsky_haurwitz(sun_apparent_elevation)

Implements the Haurwitz clear sky model for global horizontal irradiance (GHI) as presented in [1, 2].

A report on clear sky models found the Haurwitz model to have the best performance in terms of average monthly error among models which require only the Sun's elevation [26].

Arguments

sun_apparent_elevation - The apparent (refraction corrected) Sun's elevation angle. See topocentric_apparent_elevation.

Helios.clearsky_ineichen — Method

clearsky_ineichen(
    sun_apparent_elevation,
    observer_altitude,
    absolute_airmass,
    linke_turbidity,
    extraterrestial_radiation,
    perez_enhancement::Bool
)

Returns the global horizontal irradiance (GHI), direct normal irradiance (DNI), and diffuse horizontal (DHI), all in [W/m^2], following the Ineichen/Perez clear sky model [27, 28].

A report on clear sky models found the Ineichen/Perez model to have excellent performance with a minimal input data set [26].

Arguments

sun_apparent_elevation - The apparent (refraction corrected) Sun's elevation angle. See topocentric_apparent_elevation.

observer_altitude - [m] Altitude above sea level.

absolute_airmass - Pressure corrected airmass. See absolute_airmass.

linke_turbidity - Linke Turbidity. See

extraterrestial_radiation - [W/m^2] Extraterrestrial irradiance.

perez_enhancement::Bool - Controls if the Perez enhancement factor should be applied. Setting to true may produce spurious results for times when the Sun is near the horizon and the airmass is high.

Helios.clearsky_simplified_solis — Method

clearsky_simplified_solis(
    sun_apparent_elevation,
    aod700,
    precipitable_water,
    pressure,
    extraterrestial_radiation
)

Calculate the clear sky direct normal irradiance (DNI), and diffuse horizontal irradiance (DHI) according to the simplified Solis model.

Reference [29] describes the accuracy of the model as being 15, 20, and 18 W/m^2 for the beam, global, and diffuse components, respectively. Reference [30] provides comparisons with other clear sky models.

Arguments

sun_apparent_elevation - The apparent (refraction corrected) Sun's elevation angle. See topocentric_apparent_elevation.

aod700 - The aerosol optical depth at 700 nm. Algorithm derived for values between 0 and 0.45.

precipitable_water - [cm] The precipitable water of the atmosphere. Algorithm derived for values between 0.2 and 10 cm. Values less than 0.2 will be assumed to be equal to 0.2.

pressure - [Pa] The atmospheric pressure. Algorithm derived for altitudes between sea level and 7000 m, or 101325 and 41000 Pascals.

extraterrestial_radiation - [W/m^2] Extraterrestrial irradiance.