Interface

A collection of wrappers and ready-to-use utilities.

Helios.SolarPosition — Type

Represents the position of the Sun relative to an observer.

Fields

azimuth::Real: solar azimuth angle (degrees, measured clockwise from north)

elevation::Real: solar elevation angle above the horizon (degrees)

apparent_elevation::Real: elevation corrected for atmospheric refraction (degrees)

Helios.spa — Method

spa(location::Location, datetime::DateTime)

Computes the solar position in terms of Solar azimuth, elevation, and apparent elevation, following the solar position algorithm [1].