Irradiance

Helios.Irradiance — Type

Irradiance(dni, dhi, ghi)

Container for solar irradiance components.

Fields

dni: Direct Normal Irradiance [W/m²].

dhi: Diffuse Horizontal Irradiance [W/m²].

ghi: Global Horizontal Irradiance [W/m²].

Helios.angle_of_incidence_projection — Method

angle_of_incidence_projection(
    solpos::SolarPosition, surface_tilt, surface_roll, surface_azimuth
)

Calculates the dot product of the sun position unit vector and the surface normal unit vector, i.e., the cosine of the angle of incidence.

Helios.beam_component — Method

beam_component(solpos::SolarPosition, dni, surface_tilt, surface_roll, surface_azimuth)

Calculates the beam component of the plane of array irradiance.

Helios.day_angle — Method

day_angle(datetime::DateTime, offset::Int)

Calculates the day angle [degrees] for the Earth's orbit around the Sun.

For the Spencer method, offset=1; for the ASCE method, offset=0.

Helios.extraterrestrial_irradiance_asce — Method

extraterrestrial_irradiance_asce(datetime::DateTime)

Calculates the extraterrestrial irradiance [W/m²] using the ASCE evapotranspiration equation [6].

Helios.extraterrestrial_irradiance_nrel — Method

extraterrestrial_irradiance_nrel(heliocentric_radius)

Calculates the extraterrestrial irradiance [W/m²] as implemented by the NREL SPA [4].

Helios.extraterrestrial_irradiance_spencer1971 — Method

extraterrestrial_irradiance_spencer1971(datetime::DateTime)

Calculates the extraterrestrial irradiance [W/m²], as proposed in [7].

Helios.ground_diffuse_component — Method

ground_diffuse_component(ghi, albedo, surface_tilt, surface_roll)

Estimate diffuse irradiance on a tilted surface from ground reflections [8].

Helios.sky_diffuse_component_haydavies — Method

sky_diffuse_component_haydavies(
    solpos::SolarPosition, dni, dni_extra, dhi, 
    surface_tilt, surface_roll, surface_azimuth
)

Determine diffuse irradiance from the sky on a tilted surface using the Hay and Davies (1980) model [8, 9].

Helios.sky_diffuse_component_isotropic — Method

sky_diffuse_component_isotropic(dhi, surface_tilt, surface_roll)

Determine diffuse irradiance from the sky on a tilted surface using the isotropic sky model [8, 10].

Helios.sky_diffuse_component_king — Method

sky_diffuse_component_king(
    solpos::SolarPosition, dhi, ghi, surface_tilt, surface_roll, surface_azimuth
)

Determine diffuse irradiance from the sky on a tilted surface using the King model.

Helios.sky_diffuse_component_klucher — Method

sky_diffuse_component_klucher(
    solpos::SolarPosition, dhi, ghi, 
    surface_tilt, surface_roll, surface_azimuth
)

Determine diffuse irradiance from the sky on a tilted surface using the Klucher (1979) model [8, 11].

Helios.sky_diffuse_component_reindl — Method

sky_diffuse_component_reindl(
    solpos::SolarPosition, dni, dni_extra, dhi, ghi,
    surface_tilt, surface_roll, surface_azimuth
)

Determine the diffuse irradiance from the sky on a tilted surface using the Reindl (1990) model [8].

Helios.total_irradiance — Method

total_irradiance( 
    solpos::SolarPosition, irrad::Irradiance,
    surface_tilt, surface_roll, surface_azimuth;
    albedo, sky_diffuse_component
)

Determine total in-plane irradiance and its beam, sky diffuse and ground reflected components, using the specified sky diffuse irradiance model.