Solar position algorithm

The following functions implement the Solar Position Algorithm, as described in [3].

Helios.aberration_correction — Method

aberration_correction(heliocentric_radius)

An adjustment applied to celestial coordinates to account for the apparent displacement caused by Earth’s motion through space.

\[\beta = -\frac{20.4898}{3600 R_{\text{hel}}}\]

where $R_{\text{hel}}$ corresponds to the heliocentric_radius [AU].

Helios.apparent_sidereal_greenwich_time — Method

apparent_sidereal_greenwich_time(
    mean_sidereal_greenwich_time,
    nutation_longitude, 
    elliptic_obliquity
)

Greenwich sidereal time corrected for the effects of nutation giving the true rotation angle relative to the apparent equinox.

\[\tilde{t}_{\text{gw}} = \bar{t}_{\text{gw}} + \mu_{\text{nut}} \cos \omega_{\text{ell}}\]

where:

$\bar{t}_{\text{gw}}$ corresponds to mean_sidereal_greenwich_time

$\mu_{\text{nut}}$corresponds to nutation_longitude

$\omega_{\text{ell}}$corresponds to elliptic_obliquity

Helios.apparent_sun_longitude — Method

apparent_sun_longitude(geocentric_longitude, nutation_longitude, aberration_correction)

The Sun’s ecliptic longitude after applying corrections for nutation and aberration.

\[\tilde{\mu}_{\odot} = \mu_{\text{geo}} + \mu_{\text{nut}} + \beta\]

where:

$\mu_{\text{geo}}$ corresponds to geocentric_longitude

$\mu_{\text{nut}}$ corresponds to nutation_longitude

$\beta$ corresponds to aberration_correction

Helios.ascending_moon_longitude — Method

ascending_moon_longitude(julian_ephemeris_century)

The ecliptic longitude referenced to the mean equinox of the date, of the point where the Moon’s mean orbit crosses northward through the ecliptic.

\[ \mu_{\bullet} = 125.04452 - 1934.136261J_{\text{ec}} + 0.0020708J_{\text{ec}}^2 + \frac{J_{\text{ec}}^3}{450000}\]

where $J_{\text{ec}}$ is the julian_ephemeris_century.

Helios.delta_T — Method

delta_T(month::Int, year::Int)

Computes the difference between Terrestrial Dynamical Time (TD) and Universal Time (UT).

It is evaluated as a piecewise polynomial, as per NASA, which in turn refer to [4, 5]. Note that $\Delta T$ is unknown for years before -1999 and after 3000. Values could be calculated for such intervals, although they are not intended to be used for these years.

Helios.elliptic_obliquity — Method

elliptic_obliquity(mean_elliptic_obliquity, nutation_obliquity)

The angle between Earth’s equator and the mean ecliptic defined as an ideal, unperturbed ellipse.

\[\omega_{\text{ell}} = \bar{\omega}_{\text{ell}} / 3600 + \omega_{\text{nut}}\]

where:

$\bar{\omega}_{\text{ell}}$ corresponds to mean_elliptic_obliquity

$\omega_{\text{nut}}$ corresponds to nutation_obliquity

Helios.geocentric_latitude — Method

geocentric_latitude(heliocentric_latitude)

A celestial object's angular distance north or south of the ecliptic plane as measured from Earth’s center.

\[ \lambda_{\text{geo}} = -\lambda_{\text{hel}}\]

where $\lambda_{\text{hel}}$ is the heliocentric_latitude.

Helios.geocentric_longitude — Method

geocentric_longitude(heliocentric_longitude)

A celestial object's ecliptic longitude as measured from Earth’s center, referenced to the ecliptic plane and the vernal equinox.

\[\mu_{\text{geo}} = \mu_{\text{hel}} + 180 \mod 360\]

where $\mu_{\text{hel}}$ is the heliocentric_longitude, and the $\mod(\cdot)$ operator ensures that the results remains in the [0, 360] range.

Helios.geocentric_sun_ascension — Method

geocentric_sun_ascension(
    apparent_sun_longitude,
    elliptic_obliquity,
    geocentric_latitude
)

The angle measured eastward from the geocentric meridian, that locates a direction or point relative to Earth’s ellipsoid rather than its rotational axis.

\[ \sigma_{\text{geo}} = \text{atan2}(\sin\tilde{\mu}_{\odot} \cos\omega_{\text{ell}} - \tan\lambda_{\text{geo}}\sin\omega_{\text{ell}}, \cos\tilde{\mu}_{\odot})\]

where

$\tilde{\mu}_{\odot}$ corresponds to apparent_sun_longitude

$\omega_{\text{ell}}$ corresponds to elliptic_obliquity

$\lambda_{\text{geo}}$ corresponds to geocentric_latitude

Helios.geocentric_sun_declination — Method

geocentric_sun_declination(
    apparent_sun_longitude,
    elliptic_obliquity,
    geocentric_latitude
)

The Sun’s angular distance, north or south of Earth’s equatorial plane as measured from Earth’s center.

\[ \delta_{\text{geo}} = \text{asin}(\sin \lambda_{\text{geo}} \cos \omega_{\text{ell}} + \cos \lambda_{\text{geo}} \sin \omega_{\text{ell}} \sin \tilde{\mu}_{\odot}) \mod 360\]

where:

$\tilde{\mu}_{\odot}$ corresponds to apparent_sun_longitude

$\omega_{\text{ell}}$ corresponds to elliptic_obliquity

$\lambda_{\text{geo}}$ corresponds to geocentric_latitude

The $\mod(\cdot)$ operator ensures that the results remains in the [0, 360] range.

Helios.heliocentric_latitude — Method

heliocentric_latitude(julian_ephemeris_millenium)

A celestial object's angular position around the Sun measured in the ecliptic plane from a defined reference direction.

It is calculated as a polynomial expression, where the coefficients are collected in $\mathcal{L} = \{\mathbf{L}_i\}_{i=1}^2$, with $\mathbf{L}_i \in \mathbb{R}^{n_i \times 3}$ being the coeffient matrix for the $i$-th polynomal coefficient $c_{\lambda, i}$.

\[ \begin{align*} c_{\lambda, i} &= \sum_{j = 1}^{n_i} L^{(j,1)}_i \cos(L^{(j,2)}_i + L^{(j,3)}_i J_{\text{em}}) \\ \lambda_{\text{hel}} &= \frac{180}{\pi} \frac{1}{1e8} \sum_{i = 1}^2 c_{b, i} J_{\text{em}}^{i-1} \end{align*}\]

where $J_{\text{em}}$ is the julian_ephemeris_millenium. The $180/\pi$ term coverts $\lambda_{\text{hel}}$ from radians to degrees.

The coefficient matrices in $\mathcal{L}$ are collected in SolarFunctions.EARTH_PERIODIC_TERMS.L.

Helios.heliocentric_longitude — Method

heliocentric_longitude(julian_ephemeris_millenium)

A celestial object's angular distance north or south of the ecliptic plane as measured from the Sun.

It is calculated as a polynomial expression, where the coefficients are collected in $\mathcal{M} = \{\mathbf{M}_i\}_{i=1}^6$, with $\mathbf{M}_i \in \mathbb{R}^{n_i \times 3}$ being the coeffient matrix for the $i$-th polynomal coefficient $c_{\mu, i}$.

\[ \begin{align*} c_{\mu, i} &= \sum_{j = 1}^{n_i} M^{(j,1)}_i \cos(M^{(j,2)}_i + M^{(j,3)}_i J_{\text{em}}) \\ \mu_{\text{hel}} &= \frac{180}{\pi} \frac{1}{1e8} \sum_{i = 1}^6 c_{\ell, i} J_{\text{em}}^{i-1} \mod 360 \end{align*}\]

where $J_{\text{em}}$ is the julian_ephemeris_millenium. The $180/\pi$ term coverts $\mu_{\text{hel}}$ from radians to degrees.

The $\mod(\cdot)$ operator ensures that the results remains in the [0, 360] range. The coefficient matrices in $\mathcal{M}$ are collected in SolarFunctions.EARTH_PERIODIC_TERMS.M.

Helios.heliocentric_polynomial — Method

heliocentric_polynomial(julian_ephemeris_millenium, coefficients::AbstractVector)

Polynomial approximation for the heliocentric longitude, latitude, radius.

Helios.heliocentric_radius — Method

heliocentric_radius(julian_ephemeris_millenium)

The distance from the Sun to the celestial object in a chosen heliocentric coordinate system. In this case, it corresponds to the Earth-Sun distance.

It is calculated as a polynomial expression, where the coefficients are collected in $\mathcal{R} = \{\mathbf{R}_i\}_{i=1}^5$, with $\mathbf{R}_i \in \mathbb{R}^{n_i \times 3}$ being the coeffient matrix for the $i$-th polynomal coefficient $c_{R, i}$.

\[ \begin{align*} c_{R, i} &= \sum_{j = 1}^{n_i} R^{(j,1)}_i \cos(R^{(j,2)}_i + R^{(j,3)}_i J_{\text{em}}) \\ R_{\text{hel}} &= \frac{180}{\pi} \frac{1}{1e8} \sum_{i = 1}^5 c_{r, i} J_{\text{em}}^{i-1} \end{align*}\]

where $J_{\text{em}}$ is the julian_ephemeris_millenium.

The coefficient matrices in $\mathcal{R}$ are collected in SolarFunctions.EARTH_PERIODIC_TERMS.R.

Helios.julian_century — Method

julian_century(julian_day)

A time interval of exactly 36,525 days (365.25 days × 100) used as a standard unit in astronomy.

\[J_{\text{c}} = \frac{J_{\text{d}} - 2451545}{36525}\]

where $J_{\text{d}}$ is the Julian day, which can be computed using Dates.datetime2julian.

Helios.julian_ephemeris_century — Method

julian_ephemeris_century(julian_ephemeris_day)

A 36,525-day interval measured specifically in Ephemeris Time (or its modern dynamical time scales) for high-precision astronomical modeling.

\[J_{\text{ec}} = \frac{J_{\text{ed}} - 2451545}{36525}\]

where $J_{\text{ed}}$ is the julian_ephemeris_day.

Helios.julian_ephemeris_day — Method

julian_ephemeris_day(julian_day, ΔT)

A continuous count of days measured in uniform Ephemeris Time (or its successors).

\[J_{\text{ed}} = J_{\text{d}} + \frac{\Delta T}{86400}\]

where:

$J_{\text{d}}$ is the Julian day, which can be computed using Dates.datetime2julian

$\Delta T$ is the difference between Terrestrial Dynamical Time and Universal Time, which can be computed using delta_T.

Helios.julian_ephemeris_millenium — Method

julian_ephemeris_millenium(julian_ephemeris_century)

A 1,000-year interval equal to 365,250 ephemeris days, defined within Ephemeris Time for long-term astronomical calculations.

\[J_{\text{em}} = \frac{J_{\text{ec}}}{10}\]

where $J_{\text{ec}}$ is the julian_ephemeris_century.

Helios.mean_elliptic_obliquity — Method

mean_elliptic_obliquity(julian_ephemeris_millenium)

The angle between Earth’s equator and the mean (long-term averaged) ecliptic defined by an unperturbed elliptical Earth orbit.

\[ \begin{align*} U &= \frac{J_{\text{em}}}{10} \\ \bar{\omega}_{\text{ell}} &= 84381.448 − 4680.93U − 155.0U^2 + 1999.25U^3 − 51.38U^4 − \\ & \qquad 249.67U^5 − 39.05U^6 + 7.12U^7 + 27.87U^8 + 5.79U^9 + 2.45U^{10} \end{align*}\]

where $J_{\text{em}}$ is the julian_ephemeris_millenium.

Helios.mean_moon_anomaly — Method

mean_moon_anomaly(julian_ephemeris_century)

The uniformly increasing angular position of the Moon in its elliptical orbit measured from perigee.

\[ \bar{a}_{\bullet} = 134.96298 + 477198.867398J_{\text{ec}} + 0.0086972J_{\text{ec}}^2 + \frac{J_{\text{ec}}^3}{56250}\]

where $J_{\text{ec}}$ is the julian_ephemeris_century.

Helios.mean_moon_elongation_from_sun — Method

mean_moon_elongation_from_sun(julian_ephemeris_century)

The average angular separation between the Moon and the Sun as measured along the ecliptic.

\[ \bar{e}_{\bullet} = 297.85036 + 445267.111480J_{\text{ec}} − 0.0019142J_{\text{ec}}^2 + \frac{J_{\text{ec}}^3}{189474}\]

where $J_{\text{ec}}$ is the julian_ephemeris_century.

Helios.mean_sidereal_greenwich_time — Method

mean_sidereal_greenwich_time(julian_day, julian_century)

The hour angle of the mean vernal equinox at the Greenwich meridian reflecting Earth’s rotation relative to the fixed stars without nutation effects.

\[ \bar{t}_{\text{gw}} = 280.46061837 + 360.98564736629(J_{\text{d}} − 2451545.0) + 0.000387933J_{\text{c}}^2 - \frac{J_{\text{c}}^3}{38710000} \mod 360\]

where $J_{\text{d}}$ is the Julian Day, which can be computed using Dates.datetime2julian. The $\mod(\cdot)$ operator ensures that the results remains in the [0, 360] range.

Helios.mean_sun_anomaly — Method

mean_sun_anomaly(julian_ephemeris_century)

The angular position of the Sun in its elliptical orbit, measured from perihelion and increasing uniformly in time.

\[ \bar{a}_{\odot} = 357.52772 + 35999.050340J_{\text{ec}} − 0.0001603J_{\text{ec}}^2 − \frac{J_{\text{ec}}^3} {300000}\]

where $J_{\text{ec}}$ is the julian_ephemeris_century.

Helios.mean_sun_longitude — Method

mean_sun_longitude(julian_ephemeris_millenium)

The Sun’s ecliptic longitude calculated from a uniformly moving fictitious Sun on the ecliptic.

\[ \bar{\mu}_{\odot} = 280.4664567 + 360007.6982779J_{\text{em}} + 0.03032028J_{\text{em}}^2 + \frac{J_{\text{em}}^3}{49931} - \frac{J_{\text{em}}^4}{15300} - \frac{J_{\text{em}}^5}{2000000}\]

where $J_{\text{em}}$ corresponds to the julian_ephemeris_millenium.

Helios.moon_latitude_argument — Method

moon_latitude_argument(julian_ephemeris_century)

The angle from the Moon’s ascending node to its position measured along its orbit, using mean (unperturbed) orbital elements.

\[ \lambda_{\bullet} = 93.27191 + 483202.017538J_{\text{ec}} - 0.0036825J_{\text{ec}}^2 + \frac{J_{\text{ec}}^3}{327270}\]

where $J_{\text{ec}}$ is the julian_ephemeris_century.

Helios.nutation_coefficients — Method

nutation_coefficients(julian_ephemeris_century)

Constructs the vector of weights for evaluating the nutation_longitude and nutation_obliquity.

\[ \begin{align*} w_i &= \sum_{j=1}^5 X_j Y_{ij} \\ \mathbf{X} &= [\bar{e}_{\bullet}, \bar{a}_{\odot}, \bar{a}_{\bullet}, \lambda_{\bullet}, \mu_{\bullet}]^\top \end{align*}\]

The coefficients in matrix $\mathbf{Y} \in \mathbb{R}^{m \times 5}$ are collected in SolarFunctions.EARTH_PERIODIC_TERMS.Y.

Helios.nutation_longitude — Method

nutation_longitude(julian_ephemeris_century, nutation_coefficients::AbstractVector)

The small periodic variation in Earth’s ecliptic longitude caused by gravitational torques from the Moon and Sun.

\[ \mu_{\text{nut}} = \frac{1}{3.6e7} \sum_{i=1}^n (\Psi_{i,1} + \Psi_{i,2} J_{\text{ec}}) \sin w_i\]

where:

$\boldsymbol{\Psi} \in \mathbb{R}^{m \times 2}$ is the coefficient matrix, whose terms are collected in SolarFunctions.EARTH_PERIODIC_TERMS.ψ

$J_{\text{ec}}$ is the julian_ephemeris_century.

$w_i$ is the $i$-th term of the nutation_coefficients vector

Helios.nutation_obliquity — Method

nutation_obliquity(julian_ephemeris_century, nutation_coefficients::AbstractVector)

The small periodic variation in Earth’s axial tilt (obliquity), $\omega_{\text{nut}}$, resulting from gravitational perturbations.

\[ \omega_{\text{nut}} = \frac{1}{3.6e7} \sum_{i=1}^n (E_{i,1} + E_{i,2} J_{\text{ec}}) \cos w_i\]

where:

$\boldsymbol{E} \in \mathbb{R}^{m \times 2}$ is the coefficient matrix, whose terms are collected in SolarFunctions.EARTH_PERIODIC_TERMS.ϵ

$J_{\text{ec}}$ is the julian_ephemeris_century.

$w_i$ is the $i$-th term of the nutation_coefficients vector

Helios.observer_local_hour — Method

observer_local_hour(
    observer_longitude, 
    apparent_sidereal_greenwich_time, 
    geocentric_sun_ascension
)

The angle between the observer’s local meridian and a celestial object measured westward on the celestial sphere.

\[h_{\text{obs}} = \tilde{t}_{\text{gw}} + \mu_{\text{obs}} - \sigma_{\text{geo}} \mod 360\]

where:

$\mu_{\text{obs}}$ corresponds to the observer_longitude

$\tilde{t}_{\text{gw}}$ corresponds to apparent_sidereal_greenwich_time

$\sigma_{\text{geo}}$ corresponds to geocentric_sun_ascension

The $\mod(\cdot)$ operator ensures that the results remains in the [0, 360] range.

Helios.radial_distance_equatorial_plane — Method

radial_distance_equatorial_plane(
    observer_latitude,
    observer_altitude,
    reduced_observer_latitude
)

The perpendicular distance of a point from Earth’s equatorial plane.

\[d_{\text{eqt}} = \cos \lambda_{\text{red}} + A_{\text{obs}} \frac{\cos \lambda_{\text{obs}}}{6378140}\]

where:

$\lambda_{\text{obs}}$ corresponds to observer_latitude

$A_{\text{obs}}$ corresponds to observer_altitude

$\lambda_{\text{red}}$ corresponds to reduced_observer_latitude

Helios.radial_distance_rotational_axis — Method

radial_distance_rotational_axis(
    observer_latitude, 
    observer_altitude,
    reduced_observer_latitude
)

The perpendicular distance of a point from Earth’s spin axis.

\[d_{\text{rot}} = 0.99664719\sin \lambda_{\text{red}} + A_{\text{obs}} \frac{\sin \lambda_{\text{obs}}}{6378140}\]

where:

$\lambda_{\text{obs}}$ corresponds to observer_latitude

$A_{\text{obs}}$ corresponds to observer_altitude

$\lambda_{\text{red}}$ corresponds to reduced_observer_latitude

Helios.reduced_observer_latitude — Method

reduced_observer_latitude(observer_latitude)

The angle whose tangent equals the tangent of the geodetic latitude scaled by Earth’s polar-to-equatorial radius ratio, representing the point’s projection onto the surface of the reference ellipsoid.

\[\lambda_{\text{red}} = \text{atan}(0.99664719\tan(\lambda_{\text{obs}}))\]

where $\lambda_{\text{obs}}$ corresponds to the observer_latitude.

Helios.sun_ascension_parallax — Method

sun_ascension_parallax(
    radial_distance_equatorial_plane, 
    sun_equatorial_horizontal_parallax, 
    geocentric_sun_declination, 
    observer_local_hour
)

The correction to the Sun’s right ascension due to the difference between geocentric and topocentric perspectives.

\[ \xi_{\text{asc}} = \text{atan2}(-d_{\text{eqt}} \sin \xi_{\text{hor}} \sin h_{\text{obs}}, \cos \delta_{\text{geo}} - d_{\text{eqt}} \sin \xi_{\text{hor}} \cos h_{\text{obs}})\]

where:

$d_{\text{eqt}}$ corresponds to radial_distance_equatorial_plane) [m]

$\xi_{\text{hor}}$ corresponds to sun_equatorial_horizontal_parallax

$\delta_{\text{geo}}$ corresponds to geocentric_sun_declination

$h_{\text{obs}}$ corresponds to observer_local_hour

Helios.sun_equatorial_horizontal_parallax — Method

sun_equatorial_horizontal_parallax(heliocentric_radius)

The angle between the Sun’s direction as seen from Earth’s center and from a point on the equator at sea level when the Sun is on the horizon.

\[\xi_{\text{hor}} = \frac{8.794}{3600 R_{\text{hel}}}\]

where $R_{\text{hel}}$ corresponds to the heliocentric_radius, in astronomical units.

Helios.topocentric_apparent_elevation — Method

topocentric_apparent_elevation(
    observer_latitude,
    topocentric_sun_declination,
    topocentric_local_hour
)

Topocentric elevation angle without atmospheric refraction correction.

\[ \tilde{\eta}_{\text{top}} = \text{asin}(\sin \lambda_{\text{obs}} \sin \delta_{\text{top}} + \cos \lambda_{\text{obs}} \cos \delta_{\text{top}} \cos h_{\text{top}})\]

where:

$\lambda_{\text{obs}}$ corresponds to observer_latitude

$\delta_{\text{top}}$ corresponds to topocentric_sun_declination

$h_{\text{top}}$ corresponds to topocentric_local_hour

Helios.topocentric_astronomical_azimuth — Method

topocentric_astronomical_azimuth(
    observer_latitude, 
    topocentric_sun_declination, 
    topocentric_local_hour
)

The Sun’s azimuth measured from true north eastward as seen from the observer’s location, based on astronomical (not navigational) convention.

\[ \alpha_{\star, \text{top}} = \text{atan2}(\sin h_{\text{top}}, \cos h_{\text{obs}} \sin \lambda_{\text{obs}} - \tan \delta_{\text{top}} \cos \lambda_{\text{obs}}) \mod 360\]

where:

$\lambda_{\text{obs}}$ corresponds to observer_latitude

$\delta_{\text{top}}$ corresponds to topocentric_sun_declination

$h_{\text{top}}$ corresponds to topocentric_local_hour

The $\mod(\cdot)$ operator ensures that the results remains in the [0, 360] range.

Helios.topocentric_azimuth — Method

topocentric_azimuth(topocentric_astronomical_azimuth)

The Sun’s azimuth from the observer’s location relative to a defined reference direction, typically true north, on the horizon.

\[ \alpha_{\text{top}} = \alpha_{\star, \text{top}} + 180 \mod 360\]

where $\alpha_{\star, \text{top}}$ is the topocentric_astronomical_azimuth.

Helios.topocentric_elevation — Method

topocentric_elevation(
    topocentric_apparent_elevation, 
    topocentric_elevation_correction
)

The angle between the Sun and the observer’s local horizon, measured at the observer’s location.

\[\eta_{\text{top}} = \tilde{\eta}_{\text{top}} + \hat{\eta}_{\text{top}}\]

where:

$\hat{\eta}_{\text{top}}$ corresponds to the topocentric_elevation_correction

$\tilde{\eta}_{\text{top}}$ corresponds to the topocentric_apparent_elevation

Helios.topocentric_elevation_correction — Method

topocentric_elevation_correction(
    temperature,
    pressure,
    topocentric_apparent_elevation
)

The adjustment applied to an object’s observed altitude to account for the bending of its light as it passes through Earth’s atmosphere.

\[ \hat{\eta}_{\text{top}} = \frac{P}{1010} \frac{283}{273 + T} \frac{1.02}{60\tan(\tilde{\eta}_{\text{top}} + \frac{10.3}{\tilde{\eta}_{\text{top}} + 5.11})}\]

where:

$T$ corresponds to the local temperature [C]

$P$ corresponds to the local pressure [mbar]

$\tilde{\eta}_{\text{top}}$ corresponds to the topocentric_apparent_elevation

Helios.topocentric_local_hour — Method

topocentric_local_hour(observer_local_hour, sun_ascension_parallax)

The hour angle of a celestial object as seen from the observer’s exact location on Earth’s surface rather than its center.

\[h_{\text{top}} = h_{\text{obs}} - \xi_{\text{asc}}\]

where:

$h_{\text{obs}}$ corresponds to the observer_local_hour

$\xi_{\text{asc}}$ corresponds to the sun_ascension_parallax

Helios.topocentric_sun_ascension — Method

topocentric_sun_ascension(geocentric_sun_ascension, sun_ascension_parallax)

The Sun’s right ascension as viewed from the observer’s actual location on Earth’s surface.

\[\sigma_{\text{top}} = \sigma_{\text{geo}} + \xi_{\text{asc}}\]

where:

$\sigma_{\text{geo}}$ corresponds to geocentric_sun_ascension

$\xi_{\text{asc}}$ corresponds to sun_ascension_parallax

Helios.topocentric_sun_declination — Method

topocentric_sun_declination(
    radial_distance_equatorial_plane, 
    radial_distance_rotational_axis, 
    sun_equatorial_horizontal_parallax, 
    geocentric_sun_declination, 
    observer_local_hour, 
    sun_ascension_parallax
)

The Sun’s declination as viewed from the observer’s actual location on Earth’s surface.

\[ \delta_{\text{top}} = \text{atan2}((\sin \delta_{\text{geo}} - d_{\text{rot}}\sin \xi_{\text{hor}}) \cos \xi_{\text{asc}}, \cos \delta_{\text{geo}} - d_{\text{eqt}}\sin \xi_{\text{hor}} \cos h_{\text{obs}})\]

where:

$d_{\text{eqt}}$ corresponds to radial_distance_equatorial_plane [m]

$d_{\text{rot}}$ corresponds to radial_distance_rotational_axis [m]

$\xi_{\text{hor}}$ corresponds to sun_equatorial_horizontal_parallax

$\delta_{\text{geo}}$ corresponds to geocentric_sun_declination

$h_{\text{obs}}$ corresponds to observer_local_hour

$\xi_{\text{asc}}$ corresponds to sun_ascension_parallax

Helios.topocentric_sunray_direction — Method

topocentric_sunray_direction(topocentric_azimuth, topocentric_apparent_elevation)

The unit vector defining the direction of the Sun's rays, with respect to an observer on the Earth's surface. Conventionally, it points from the observer towards the Sun [6].

\[ \hat{\mathbf{s}} = \begin{bmatrix} \sin \alpha_{\text{top}} \cos \tilde{\eta}_{\text{top}} \\ \cos \alpha_{\text{top}} \cos \tilde{\eta}_{\text{top}} \\ \sin \tilde{\eta}_{\text{top}} \end{bmatrix}\]

where:

$\tilde{\eta}_{\text{top}}$ corresponds to the topocentric_apparent_elevation

$\alpha_{\text{top}}$ corresponds to the topocentric_azimuth