Atmosphere

Helios.absolute_airmass — Method

absolute_airmass(relative_airmass, pressure)

Computes the absolute airmass [5].

Helios.altitude2pressure — Method

altitude2pressure(altitude)

Computes the site pressure from local altitude.

It follows the same assumption of pressure2altitude.

Helios.aod_bb_hulstrom1980 — Method

aod_bb_hulstrom1980(aod380, aod500)

Computes the approximate broadband aerosol optical depth (AOD), as proposed in [6, 7].

It correlates broadband AOD using two wavelengths, 380 nm and 500 nm, denoted with aod380 and aod500, respectively.

Helios.apparent_vapor_scale_height_gueymard94 — Method

apparent_vapor_scale_height_gueymard94(air_temperature)

Computes the apparent water vapor scale height [km] as proposed in [8, Eq. 4].

Helios.linke_turbidity_kasten1996 — Method

linke_turbidity_kasten1996(absolute_airmass, precipitable_water, aod_bb)

Computes the Linke turbidity using Kasten pyrheliometric formula.

Note that broadband aerosol optical depth (AOD) can be approximated by AOD measured at 700 nm according to Molineaux [9]. Bird and Hulstrom offer an alternate approximation using broadband aerosol optical depth (aod_bb), measured at 380 nm and 500 nm. Based on original implementation by Armel Oumbe. Note that these calculations are only valid for absolute_airmass less than 5 and precipitable_water less than 5 cm.

Helios.linke_turbidity_meteotest — Method

linke_turbidity_meteotest(location::Location, datetime::DateTime)

Computes the Linke turbidity based on SoDa [10, 11].

It performs a bilinear interpolation of the Linke turbidity array, for the corrsponding latitude, longitude, and month values. In such an array, rows represent global latitudes, columns represent global longitudes, and depth (third dimension) represents months of the year.

Helios.precipitable_water_gueymard94 — Method

precipitable_water_gueymard94(saturation_vapor_pressure, surface_vapor_density)

Computes precipitable water [cm] as proposed in [8, Eq. 1].

The accuracy of this method is approximately 20% for moderate precipitable water (1-3 cm) and less accurate otherwise.

Helios.pressure2altitude — Method

pressure2altitude(pressure)

Computes the local altitude from site pressure.

The following assumptions [12] hold:

Base pressure: 101325 Pa

Temperature at zero altitude: 288.15 K

Gravitational acceleration: 9.80665 m/s^2

Lapse rate: -6.5E-3 K/m

Gas constant for air: 287.053 J/(kg K)

Relative Humidity: 0%

Helios.relative_airmass_gueymard1993 — Method

relative_airmass_gueymard1993(solpos::SolarPosition)

Computes the relative airmass as proposed in [5, 13].

Helios.relative_airmass_gueymard2003 — Method

relative_airmass_gueymard2003(solpos::SolarPosition)

Computes the relative airmass as proposed in [14, 15].

Helios.relative_airmass_kasten1966 — Method

relative_airmass_kasten1966(solpos::SolarPosition)

Computes the relative airmass as proposed in [16].

Helios.relative_airmass_kastenyoung1989 — Method

relative_airmass_kastenyoung1989(solpos::SolarPosition)

Computes the relative airmass as proposed in [17].

Helios.relative_airmass_pickering2002 — Method

relative_airmass_pickering2002(solpos::SolarPosition)

Computes the relative airmass as proposed in [18].

Helios.relative_airmass_simple — Method

relative_airmass_simple(solpos::SolarPosition)

Computes the relative airmass as secant of the Sun's apparent zenith.

Helios.relative_airmass_young1994 — Method

relative_airmass_young1994(solpos::SolarPosition)

Computes the relative airmass as proposed in [19].

Helios.relative_airmass_youngirvine1967 — Method

relative_airmass_youngirvine1967(solpos::SolarPosition)

Computes the relative airmass as proposed in [20].

Helios.relative_humidity_oke2018 — Method

relative_humidity_oke2018(vapor_pressure, saturation_vapor_pressure)

Computes the relative humidity [%] as proposed in [21].

Helios.saturation_vapor_pressure_gueymard93 — Method

saturation_vapor_pressure_gueymard93(air_temperature)

Computes the saturation water vapor pressure [mbar] as proposed in [5, Eq. 1].

Helios.saturation_vapor_pressure_oke2018 — Method

saturation_vapor_pressure_oke2018(dew_temperature)

Computes the saturation water vapor pressure [mbar] as proposed in [21].

Helios.surface_vapor_density_gueymard94 — Method

surface_vapor_density_gueymard94(
    air_temperature,
    relative_humidity,
    apparent_vapor_scale_height
)

Computes the surface water vapor density [g/m^3] as proposed in [8, Eq. 2].

Helios.vapor_pressure_oke2018 — Method

vapor_pressure_oke2018(air_temperature)

Computes the water vapor pressure [mbar] as proposed in [21].