Interface

A collection of wrappers and ready-to-use utilities.

Helios.Irradiance — Type

Irradiance(dni, dhi, ghi)

Container for solar irradiance components.

Arguments

dni: Direct Normal Irradiance (W/m²).

dhi: Diffuse Horizontal Irradiance (W/m²).

ghi: Global Horizontal Irradiance (W/m²).

Helios.Location — Type

Represents a physical geographic location together with ambient conditions.

Helios.Location — Method

Location(latitude, longitude, altitude, temperature=25.0, pressure=nothing)

Create a Location from geographic coordinates and ambient conditions.

Arguments

latitude: latitude (degrees)

longitude: longitude (degrees)

altitude: altitude above sea level (meters)

temperature: ambient temperature (°C)

Keyword

pressure: ambient pressure; if not provided, it is calculated using

altitude2pressure.

Helios.SolarPosition — Type

Represents the position of the Sun relative to an observer.

Fields

azimuth::Real: solar azimuth angle (degrees, measured clockwise from north)

elevation::Real: solar elevation angle above the horizon (degrees)

apparent_elevation::Real: elevation corrected for atmospheric refraction (degrees)

Helios.clearsky_haurwitz — Method

clearsky_haurwitz(solpos::SolarPosition)

Computes the GHI irradiance component following the Haurwitz model [1, 2]. It is a wrapper for clearsky_haurwitz.

Helios.clearsky_ineichen — Method

clearsky_ineichen(
    location::Location, 
    solpos::SolarPosition, 
    datetime::DateTime;
    func_relative_airmass=relative_airmass_kastenyoung1989,
    perez_enhancement = false
)
clearsky_ineichen(
    location::Location, 
    datetime::DateTime;
    func_relative_airmass=relative_airmass_kastenyoung1989,
    perez_enhancement = false
)

Computes the DNI, DHI, and GHI irradiance components following the Ineichen model [ineichen2002new, perez2002new]. It is a wrapper for clearsky_ineichen.

Helios.spa — Method

spa(location::Location, datetime::DateTime)

Computes the solar position in terms of Solar azimuth, elevation, and apparent elevation, following the solar position algorithm [3].

Helios.sunray — Method

sunray(solpos::SolarPosition)

The unit vector defining the direction of the Sun's rays.